refactor:  Gambler의 추상 클래스 해제
